@@ -31,54 +31,55 @@
     "@testing-library/dom": "^9.3.1",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^14.0.0",
-    "@testing-library/react-hooks": "^8.0.1",
     "@testing-library/user-event": "^14.4.3",
-    "@types/jest": "^26.0.23",
+    "@types/jest": "^29.5.3",
     "@types/react-router-dom": "^5.3.3",
-    "@types/victory": "^33.1.5",
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
-    "@webpack-cli/serve": "^1.7.0",
-    "copy-webpack-plugin": "^7.0.0",
+    "@types/victory": "^35.0.0",
+    "@typescript-eslint/eslint-plugin": "^6.3.0",
+    "@typescript-eslint/parser": "^6.3.0",
+    "@webpack-cli/serve": "^2.0.5",
+    "ajv": "^8.12.0",
+    "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^5.2.6",
-    "css-minimizer-webpack-plugin": "^3.4.1",
-    "cypress": "^11.2.0",
-    "dotenv-webpack": "^7.1.0",
-    "eslint": "^7.32.0",
-    "eslint-plugin-i18n-json": "^3.1.0",
-    "eslint-plugin-jest-dom": "^3.9.4",
-    "eslint-plugin-react": "^7.30.0",
+    "css-minimizer-webpack-plugin": "^5.0.1",
+    "cypress": "^12.17.3",
+    "dotenv-webpack": "^8.0.1",
+    "eslint": "^8.46.0",
+    "eslint-plugin-i18n-json": "^4.0.0",
+    "eslint-plugin-jest-dom": "^5.0.2",
+    "eslint-plugin-react": "^7.33.1",
     "eslint-plugin-react-hooks": "^4.6.0",
     "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.5.0",
+    "html-webpack-plugin": "^5.5.3",
     "imagemin": "^8.0.1",
-    "jest": "^27.5.1",
+    "jest": "^29.6.2",
+    "jest-environment-jsdom": "^29.6.2",
     "license-reporter": "1.5.0",
-    "mini-css-extract-plugin": "^1.6.2",
-    "monaco-editor-webpack-plugin": "^2.1.0",
-    "prettier": "^2.7.1",
+    "mini-css-extract-plugin": "^2.7.6",
+    "monaco-editor-webpack-plugin": "^7.1.0",
+    "prettier": "^3.0.1",
     "prop-types": "^15.8.1",
     "raw-loader": "^4.0.2",
     "react-axe": "^3.5.4",
     "react-router-dom": "^5.3.3",
-    "react-test-renderer": "17.0.2",
-    "regenerator-runtime": "^0.13.9",
-    "rimraf": "^3.0.2",
-    "start-server-and-test": "^1.14.0",
-    "style-loader": "^2.0.0",
-    "svg-url-loader": "^7.1.1",
-    "terser-webpack-plugin": "^5.3.3",
-    "ts-jest": "^27.1.5",
-    "ts-loader": "^9.3.0",
-    "tsconfig-paths-webpack-plugin": "^3.5.2",
-    "tslib": "^2.4.0",
-    "typescript": "^4.7.3",
+    "react-test-renderer": "18.2.0",
+    "regenerator-runtime": "^0.14.0",
+    "rimraf": "^5.0.1",
+    "start-server-and-test": "^2.0.0",
+    "style-loader": "^3.3.3",
+    "svg-url-loader": "^8.0.0",
+    "terser-webpack-plugin": "^5.3.9",
+    "ts-jest": "^29.1.1",
+    "ts-loader": "^9.4.4",
+    "tsconfig-paths-webpack-plugin": "^4.1.0",
+    "tslib": "^2.6.1",
+    "typescript": "^5.1.6",
     "url-loader": "^4.1.1",
-    "webpack": "^5.38.1",
-    "webpack-bundle-analyzer": "^4.4.2",
-    "webpack-cli": "^5.0.1",
-    "webpack-dev-server": "^4.11.1",
-    "webpack-merge": "^5.8.0"
+    "webpack": "^5.88.2",
+    "webpack-bundle-analyzer": "^4.9.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1",
+    "webpack-merge": "^5.9.0"
   },
   "dependencies": {
     "@patternfly/react-charts": "^7.0.0",
