@@ -1,12 +1,6 @@
 import React, { useEffect, useState } from 'react';
-import {
-	TableVariant
-} from '@patternfly/react-table';
-import {
-	Table,
-	TableBody,
-	TableHeader
-} from '@patternfly/react-table/deprecated';
+import { TableVariant } from '@patternfly/react-table';
+import { Table, TableBody, TableHeader } from '@patternfly/react-table/deprecated';
 import {
   Badge,
   Bullseye,
@@ -21,8 +15,8 @@ import {
   StackItem,
   Text,
   TextContent,
-  TextVariants, EmptyStateHeader,
-  
+  TextVariants,
+  EmptyStateHeader
 } from '@patternfly/react-core';
 import { SearchIcon } from '@patternfly/react-icons';
 import displayUtils from '@services/displayUtils';
@@ -156,7 +150,11 @@ const TasksTableDisplay = (props: { setTasksCount: (number) => void; isVisible: 
               title: (
                 <Bullseye>
                   <EmptyState variant={EmptyStateVariant.sm}>
-                    <EmptyStateHeader titleText={<>{t('cache-managers.no-tasks-status')}</>} icon={<EmptyStateIcon icon={SearchIcon} />} headingLevel="h2" />
+                    <EmptyStateHeader
+                      titleText={<>{t('cache-managers.no-tasks-status')}</>}
+                      icon={<EmptyStateIcon icon={SearchIcon} />}
+                      headingLevel="h2"
+                    />
                     <EmptyStateBody>{t('cache-managers.no-tasks-body')}</EmptyStateBody>
                   </EmptyState>
                 </Bullseye>
