@@ -1,16 +1,6 @@
 import React, { useEffect, useState } from 'react';
-import {
-	Alert,
-	Form,
-	FormAlert,
-	FormGroup,
-	FormSection
-} from '@patternfly/react-core';
-import {
-	Select,
-	SelectOption,
-	SelectVariant
-} from '@patternfly/react-core/deprecated';
+import { Alert, Form, FormAlert, FormGroup, FormSection } from '@patternfly/react-core';
+import { Select, SelectOption, SelectVariant } from '@patternfly/react-core/deprecated';
 import { CacheFeature, CacheMode } from '@services/infinispanRefData';
 import { useTranslation } from 'react-i18next';
 import { ConsoleServices } from '@services/ConsoleServices';
