@@ -20,7 +20,8 @@ import {
   ToolbarItemVariant
 } from '@patternfly/react-core';
 import { Select, SelectOption, SelectVariant } from '@patternfly/react-core/deprecated';
-import { Table, TableVariant, Tbody, Td, Th, Thead, Tr } from '@patternfly/react-table';
+import { TableVariant, Tbody, Td, Th, Thead, Tr } from '@patternfly/react-table';
+import { Table } from '@patternfly/react-table/deprecated';
 import { Chart, ChartBar, ChartGroup, ChartVoronoiContainer, ChartThemeColor } from '@patternfly/react-charts';
 import { SearchIcon } from '@patternfly/react-icons';
 import { TableErrorState } from '@app/Common/TableErrorState';
