@@ -26,7 +26,8 @@ import {
   EmptyStateHeader
 } from '@patternfly/react-core';
 import { CubesIcon, SearchIcon, DownloadIcon } from '@patternfly/react-icons';
-import { Table, Thead, Tr, Th, Tbody, Td } from '@patternfly/react-table';
+import { Thead, Tr, Th, Tbody, Td } from '@patternfly/react-table';
+import { Table } from '@patternfly/react-table/deprecated';
 import { Health } from '@app/Common/Health';
 import { TableErrorState } from '@app/Common/TableErrorState';
 import { useTranslation } from 'react-i18next';
