@@ -6,7 +6,7 @@ import i18n from './i18n4Test';
 import { I18nextProvider } from 'react-i18next';
 
 export function renderWithRouter(ui, { route = '/', history = createMemoryHistory({ initialEntries: [route] }) } = {}) {
-  const Wrapper: FunctionComponent = ({ children }) => (
+  const Wrapper = ({ children }) => (
     <Router history={history}>
       <I18nextProvider i18n={i18n}>{children}</I18nextProvider>
     </Router>
